order-service/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
sqlalchemy>=2.0
psycopg2-binary>=2.9
alembic>=1.13
pydantic>=2.6
httpx>=0.27
aio-pika>=9.0
python-jose[cryptography]>=3.3

# Test
pytest>=8.0
pytest-asyncio>=0.23
prometheus-fastapi-instrumentator>=6.1
